Create GetPlayerID.scpt
File is an Applescript file that will get Opponent ID after the game started and you press OK button. It will get 7 digits of player ID, so if player ID has less numbers it still needs to be trimmed, but I did not work on this yet. In the end it will try to Open GUDecks website with the opponent player ID.